"Refactor methods to improve testability"
In 'OpenSearchSink.cs', the accessibility of 'OnProcessExit' method was modified from 'private' to 'internal'. This change allows the method to be tested. Also, in 'TestFixture.cs' and 'OpenSearchSinkTests.cs', there was a refactoring where the 'GetLogger' method now returns an OpenSearchSink instance instead of OpenSearchClient. 'OpenSearchSink' provides more access to internal state and operations. This will fortify our tests as it allows for an increased range of assertions and interactions. Lastly, test cases were extended to increase coverage and ensure all log sending scenarios work as expected.