Switch from ConcurrentQueue to BlockingCollection for logs queue
Replacing the ConcurrentQueue with a BlockingCollection for the logs queue adds an optional limit on the queue size. This switch allows for more control over memory management, as we can now limit the maximum number of logs in the queue, dropping new logs when this limit is reached. These changes implement the previously considered queue bounding like in Grafana Loki Sink.